Chap06/SpeechRecognition/SpeechRecognition/requirements.txt
absl-py==0.9.0
adal==1.2.5
antlr4-python3-runtime==4.7.2
appdirs==1.4.4
applicationinsights==0.11.9
argcomplete==1.12.2
astor==0.7.1
astunparse==1.6.3
attrs==19.1.0
audioread==2.1.9
azure-ai-textanalytics==5.1.0
azure-appconfiguration==1.1.1
azure-batch==9.0.0
azure-cli==2.16.0
azure-cli-core==2.16.0
azure-cli-telemetry==1.0.6
azure-cognitiveservices-speech==1.14.0
azure-common==1.1.26
azure-core==1.17.0
azure-cosmos==3.2.0
azure-datalake-store==0.0.51
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b2
azure-loganalytics==0.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.2.0
azure-mgmt-appconfiguration==0.6.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==9.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==1.0.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==5.2.0
azure-mgmt-cognitiveservices==6.3.0
azure-mgmt-compute==14.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc15
azure-mgmt-containerservice==9.4.0
azure-mgmt-core==1.2.1
azure-mgmt-cosmosdb==1.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==3.0.0rc7
azure-mgmt-eventhub==4.1.0
azure-mgmt-hdinsight==2.0.0
azure-mgmt-imagebuilder==0.4.0
azure-mgmt-iotcentral==3.0.0
azure-mgmt-iothub==0.12.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==8.0.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.7.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==2.2.0
azure-mgmt-monitor==0.11.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.14.0
azure-mgmt-network==13.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.5.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==3.1.0rc1
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.6.0
azure-mgmt-redhatopenshift==0.1.0
azure-mgmt-redis==7.0.0rc2
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==10.3.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.4.1
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.5.0
azure-mgmt-signalr==0.4.0
azure-mgmt-sql==0.21.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==16.0.0
azure-mgmt-synapse==0.5.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==0.48.0
azure-multiapi-storage==0.5.2
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.2.0
azure-synapse-artifacts==0.3.0
azure-synapse-spark==0.2.0
backcall==0.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bleach==1.5.0
branca==0.3.1
bs4==0.0.1
cachetools==4.0.0
certifi==2018.10.15
cffi==1.14.4
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
colorama==0.4.1
colorlover==0.3.0
combo==0.1.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.7
dataclasses==0.8
DateTime==4.3
decorator==4.4.0
deepmachine==0.7.5
defusedxml==0.6.0
dill==0.3.1.1
easyocr==1.1.3
entrypoints==0.3
enum34==1.1.10
fabric==2.5.0
Faker==3.0.0
Flask==1.1.1
Flask-Cors==3.0.8
folium==0.10.1
future==0.18.2
gast==0.3.3
googleapis-common-protos==1.51.0
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
humanfriendly==8.2
idna==2.6
imageio==2.5.0
imbalanced-learn==0.6.2
imblearn==0.0
importlib-metadata==1.5.0
imutils==0.5.3
invoke==1.4.1
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
itsdangerous==1.1.0
javaproperties==0.5.1
jedi==0.13.3
Jinja2==2.11.1
jmespath==0.10.0
joblib==1.0.1
jsmin==2.2.2
jsondiff==1.2.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.0.2
kiwisolver==1.1.0
knack==0.7.2
librosa==0.8.0
lightgbm==2.3.1
llvmlite==0.31.0
lxml==4.5.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
mongoengine==0.20.0
mpmath==1.1.0
msal==1.0.0
msal-extensions==0.1.3
msrest==0.6.21
msrestazure==0.6.4
mtcnn==0.1.0
mxnet==1.3.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nibabel==3.2.1
notebook==5.7.8
numba==0.48.0
numpy==1.18.1
oauthlib==3.1.0
onnx==1.6.0
OpenCC==1.1.1
opencv-contrib-python==4.5.2.54
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.9
pandas==0.25.1
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.4.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.0.0
pip==21.1.1
pkginfo==1.6.1
plotly==2.7.0
pooch==1.3.0
portalocker==1.7.1
prometheus-client==0.6.0
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.11.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydot==1.4.1
Pygments==2.4.0
PyJWT==2.0.0
pymongo==3.11.0
PyNaCl==1.4.0
pyod==0.7.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.15.2
pytesseract==0.3.4
python-dateutil==2.8.0
pytorch-ranger==0.1.1
pytz==2019.1
PyWavelets==1.0.3
pywin32==227
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
resnet==0.1
retrying==1.3.3
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.24.0
scikit-optimize==0.5.2
scipy==1.4.1
scp==0.13.3
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
setuptools==45.2.0
six==1.12.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
SpeechRecognition==3.8.1
sshtunnel==0.1.5
statsmodels==0.12.1
suod==0.0.3
sympy==1.5.1
tabulate==0.8.7
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.3.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.3.0
tensorflow-gpu-estimator==2.1.0
tensorflow-metadata==0.21.1
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
text-unidecode==1.3
threadpoolctl==2.1.0
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tornado==6.0.2
tqdm==4.43.0
traitlets==4.3.2
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.22
vsts==0.1.25
vsts-cd-manager==1.0.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.0
wheel==0.32.3
widgetsnbextension==3.4.2
wrapt==1.12.0
xgboost==0.90
xlrd==1.2.0
xmltodict==0.12.0
zipp==3.0.0
zope.interface==4.6.0
